Fixed a few warnings. Now compiles with OS 10.10. History write is inconsistent and display font doesn't resize with window changes.

+++ Constant.m
@@ -71,7 +71,7 @@
 {
 	[super encodeWithCoder:coder];
 
-	[coder encodeInt:constant forKey:@"MEConstant"];
+	[coder encodeInt:(int)constant forKey:@"MEConstant"];
 	[coder encodeObject:value forKey:@"MEValue"];
 }
 
@@ -117,7 +117,7 @@
 {
 	// Behave as though we already have a child and spawn off another node
 	[self binaryOpPressed:'.'];
-	[[manager getInputPoint] constantPressed:newConstant];
+	[[manager getInputPoint] constantPressed:(int)newConstant];
 }
 
 //
